+++ src/com/genir/aitweaks/core/features/shipai/WeaponGroup.kt
@@ -3,11 +3,12 @@ package com.genir.aitweaks.core.features.shipai
 import com.fs.starfarer.api.combat.ShipAPI
 import com.fs.starfarer.api.combat.WeaponAPI
 import com.genir.aitweaks.core.utils.extensions.isAngleInArc
+import com.genir.aitweaks.core.utils.extensions.sumOf
 
 /** A group of weapons that can focus on a single attack angle. */
 class WeaponGroup(val ship: ShipAPI, significantWeapons: List<WeaponAPI>, val facing: Float) {
     val weapons = significantWeapons.filter { it.isAngleInArc(facing) }
-    val dps: Float = weapons.sumOf { it.derivedStats.dps.toDouble() }.toFloat()
+    val dps: Float = weapons.sumOf { it.derivedStats.dps }
     val effectiveRange: Float = effectiveRange(Preset.effectiveDpsThreshold)
     val minRange: Float = weapons.minOfOrNull { it.slotRange } ?: 0f
     val maxRange: Float = weapons.maxOfOrNull { it.slotRange } ?: 0f
